+++ utitool/cli.swift
@@ -64,4 +64,20 @@ struct Cli {
 
         return newArgs
     }
+
+
+    /**
+        Get the value of a named shell environment variable.
+
+        - Parameters
+            - envVar: The environment variable, eg. `TERM`.
+
+        - Returns The environment variable's value as a string,
+                  or an empty string on error/absence.
+     */
+    static func getEnvVar(_ envVar: String) -> String {
+
+        guard let rawValue = getenv(envVar) else { return "" }
+        return String(utf8String: rawValue) ?? ""
+    }
 }